initial commit of midi with device/port
