+++ .gitignore
@@ -110,3 +110,5 @@ $RECYCLE.BIN/
 # Mac crap
 .DS_Store
 
+# Xcode project
+xcode/
