+++ firmware/usbh_conf.c
@@ -42,6 +42,18 @@ HCD_HandleTypeDef hHCD;
 #include "usbh_midi_core.h"
 #include "ch.h"
 
+
+#include "midi.h"
+
+//extern void MidiInMsgHandler(uint8_t status, uint8_t data1, uint8_t data2);
+
+//TODO: need incoming port number
+void MIDI_CB(uint8_t a,uint8_t b,uint8_t c,uint8_t d){
+    USBH_DbgLog("M %x - %x %x %x\r\n",a,b,c,d);
+    //  a= pkt header 0xF0 = cable number 0x0F=CIN
+    MidiInMsgHandler(MIDI_DEVICE_USB_HOST, ((a & 0xF0) >> 4)+ 1 ,b,c,d);
+}
+
 USBH_HandleTypeDef hUSBHost; /* USB Host handle */
 static void USBH_UserProcess(USBH_HandleTypeDef *pHost, uint8_t vId);
 
@@ -571,12 +583,6 @@ void USBH_HID_EventCallback(USBH_HandleTypeDef *phost) {
   }
 }
 
-extern void MidiInMsgHandler(uint8_t status, uint8_t data1, uint8_t data2);
-
-void MIDI_CB(uint8_t a,uint8_t b,uint8_t c,uint8_t d){
-  USBH_DbgLog("M %x %x %x %x\r\n",a,b,c,d);
-  MidiInMsgHandler(b,c,d);
-}
 
 #define PORT_IRQ_HANDLER(id) void id(void)
 #define CH_IRQ_HANDLER(id) PORT_IRQ_HANDLER(id)
