+++ firmware/usbh_midi_core.c
@@ -41,6 +41,28 @@
 /* Includes ------------------------------------------------------------------*/
 #include "usbh_midi_core.h"
 
+
+void usbh_midi_init(void)
+{
+    
+}
+
+void usbh_MidiSend1(uint8_t port, uint8_t b0)
+{
+    
+}
+
+void usbh_MidiSend2(uint8_t port, uint8_t b0, uint8_t b1)
+{
+    
+}
+
+void usbh_MidiSend3(uint8_t port, uint8_t b0, uint8_t b1, uint8_t b2)
+{
+    
+}
+
+
 /** @defgroup USBH_MIDI_CORE_Private_Variables
  * @{
  */
