+++ firmware/usbh_midi_core.h
@@ -34,6 +34,15 @@
 //#include "usbh_usr.h"
 //#include "midi_interface.h"
 
+
+// external midi interface
+void usbh_midi_init(void);
+void usbh_MidiSend1(uint8_t port, uint8_t b0);
+void usbh_MidiSend2(uint8_t port, uint8_t b0, uint8_t b1);
+void usbh_MidiSend3(uint8_t port, uint8_t b0, uint8_t b1, uint8_t b2);
+
+
+
 //#define MIDI_MIN_POLL          10
 #define USBH_MIDI_MPS_SIZE  64
 #define USB_AUDIO_CLASS 0x01
